Changes to v4 for debugging purposes
I've made several changes to v4 to allow us to debug which random effect
is causing problems.  I've convinced myself that it works with all
random effects turned off, so now we add them back on at a time, or in
combination, to figure out which is causing the problem.